+++ OnlineVideoLinks/Utilities/GameVideoUtility.cs
@@ -139,7 +139,7 @@ namespace OnlineVideoLinks.Utilities
                     }
 
                     // TEMP
-                    if (app.Name == GameVideo.GamesDbVideoTitleWithPrefix)
+                    if (app.Name == GameVideo.GamesDbVideoTitleWithPrefix || app.Name == GameVideo.GamesDbVideoTitleNoPrefix)
                         game.TryRemoveAdditionalApplication(app);
                 }
             }
